requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
sqlparse>=0.4.0
googletrans-py>=1.1.2
pandas>=1.3.0

